third_party/blink/web_tests/platform/generic/http/tests/devtools/network/json-preview-expected.txt

Tests that resources with JSON MIME types are previewed with the JSON viewer.


Running: test1
Creating a NetworkRequest with type: application/json
Its previewer is searchable: true
Its previewer is the JSON previewer: true

Running: test2
Creating a NetworkRequest with type: application/vnd.document+json
Its previewer is searchable: true
Its previewer is the JSON previewer: true
